requirements.txt
# Brain Hemorrhage Segmentation - Requirements
# Resolved dependency conflicts

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Medical Imaging
monai>=1.3.0
nibabel>=5.0.0
pydicom>=2.4.0

# Scientific Computing (Compatible versions)
numpy>=2.0.0
pandas>=2.2.2
scikit-learn==1.7.1

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Flask API
flask>=3.0.0
werkzeug>=3.0.0
flask-cors>=4.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.14.0

# Optional: For visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For DICOM conversion
SimpleITK>=2.3.0